+++ src/collections/AutoIncrementIdentifierSet.ts
@@ -1,0 +1,10 @@
+export class AutoIncrementIdentifierSet extends Set<number> {
+  #minValue = 0;
+  nextValue() {
+    let value = this.#minValue;
+    while (this.has(value)) {
+      value++;
+    }
+    return value;
+  }
+}
