+++ src/collections/index.ts
@@ -1,0 +1,7 @@
+/**
+ *
+ * TODO move collections from ./Observable to here and create an Observable mixin/decorator
+ */
+export { AutoIncrementIdentifierSet } from "./AutoIncrementIdentifierSet";
+export { NumberKeyedMap } from "./NumberKeyedMap";
+export { InstanceMap } from "./InstanceMap";
